api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
tensorflow==2.20.0
pillow==10.4.0
numpy==1.26.4
opencv-python==4.10.0.84
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==23.2.1
kaggle==1.6.17
